requirements.txt
ollama
tavily-python
python-dotenv
chromadb